requirements.txt
cloudpickle==2.2.1
entrypoints==0.4
gym==0.21.0
gym-notices==0.0.8
gym-retro==0.8.0
importlib-metadata==4.13.0
numpy==1.21.6
opencv-python==4.11.0.86
pyglet==1.5.27
typing_extensions==4.7.1
zipp==3.15.0
